okay

+++ back-end/doctormodule/src/main/java/com/medicalservice/doctor/controller/DoctorController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
+//import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -16,28 +16,31 @@ import com.medicalservice.doctor.service.DoctorServiceImp;
 
 @RestController
 public class DoctorController {
-	
+
 	@Autowired
 	DoctorServiceImp doctorServiceImp;
-	
+
 	@GetMapping("/getDoctors")
-	public List<Doctor> getDoctors(){
+	public List<Doctor> getDoctors() {
 		return doctorServiceImp.getDoctors();
 	}
+
 	@PostMapping("/addDoctor")
 	public void addDoctor(@RequestBody Doctor doctor) {
 		doctorServiceImp.addDoctor(doctor);
 	}
+
 //	@PutMapping("/updateDoctor")
 //	public void updateDoctor(@RequestBody Doctor doctor){
 //		doctorServiceImp.updateDoctor(doctor);
 //	}
 	@GetMapping("/findDoctorById")
-	public Doctor findDoctorById(@PathVariable int id) {
+	public Doctor findDoctorById(@PathVariable Long id) {
 		return doctorServiceImp.findDoctorById(id);
 	}
+
 	@DeleteMapping("/deleteDoctorById")
-	public void deleteDoctorById(@PathVariable int id) {
+	public void deleteDoctorById(@PathVariable Long id) {
 		doctorServiceImp.deleteDoctorById(id);
 	}
